requirements.txt
slumber==0.7.0

# Testing
coverage==3.7.1
ddt==1.0.0
httpretty==0.8.8
mock==1.0.1
nose==1.3.6
pep8==1.6.2
PyJWT==1.1.0
pylint==1.4.3
